Move supporting functions from orm_database.py to database.py
